Just fixing some errors

+++ src/Model/movingobjects.cpp
@@ -12,9 +12,9 @@ void MovingObject::move() {
 }
 
 bool MovingObject::collide(Frog& frog) {
-    return ((getUpX() <= frog.getX() && getUpX() + getSize() >= frog.getX()) ||
-            (getUpX() <= frog.getX() + constants::frog::WIDTH && getUpX()
-             + getSize() >= frog.getX() + constants::frog::WIDTH)) &&
+    return ((static_cast<int>(getUpX()) <= frog.getX() && static_cast<int>(getUpX() + getSize()) >= frog.getX()) ||
+            (static_cast<int>(getUpX()) <= frog.getX() + constants::frog::WIDTH && static_cast<int>(getUpX()
+             + getSize()) >= frog.getX() + constants::frog::WIDTH)) &&
         frog.getLane() == getId();
 }
